Patch 5
I got most of it working, but there are still some problems that need ironing out. I switched gears from using a dynamic array to a fixed one to get around a problem, but their is still some work. I included my backup of that work if you guys want to have a go.